store2bucket_function/requirements.txt
google-cloud-pubsub==1.4.2
google-cloud-logging
google-cloud-storage==1.28.0
pymysql